git_log function in new file git_log.py
