+++ setup.py
@@ -11,8 +11,8 @@ setup(
     packages=find_packages("repo_sync"),
     entry_points={
         'console_scripts': [
-            'repo-sync=repo_sync:sync',
-            'repo-sync=repo_sync:git_log',
+            'repo-sync1=repo_sync.cli:sync',
+            'repo-sync2=repo_sync.git_log:git_log',
         ],
     },
 )
